feat: add token validation on deleteCategory

+++ src/features/posts/actions/deleteCategory.ts
@@ -2,10 +2,28 @@
 
 import { connectDatabase } from "@/lib/db";
 import { Category } from "../models/Category";
+import jwt from "jsonwebtoken";
+import { cookies } from "next/headers";
+
+const SECRET_KEY = process.env.JWT_SECRET!;
+
 
 export async function deleteCategory(id: string) {
   try {
     await connectDatabase();
+
+    const cookieStore = await cookies();
+    const token = cookieStore.get("token")?.value;
+
+    if (!token) {
+      return { success: false, message: "Usuário não autenticado." };
+    }
+
+    try {
+          jwt.verify(token, SECRET_KEY);
+    } catch {
+          return { success: false, message: "Token inválido." };
+      }
 
     const deleted = await Category.findByIdAndDelete(id).lean();
